+++ src/main/java/net/laboulangerie/laboulangeriemmo/LaBoulangerieMmo.java
@@ -13,15 +13,15 @@ import net.laboulangerie.laboulangeriemmo.api.player.MmoPlayerManager;
 import net.laboulangerie.laboulangeriemmo.api.talent.TalentsRegistry;
 import net.laboulangerie.laboulangeriemmo.betonquest.LevelCondition;
 import net.laboulangerie.laboulangeriemmo.betonquest.XpEvent;
-import net.laboulangerie.laboulangeriemmo.blockus.BlockusDataManager;
-import net.laboulangerie.laboulangeriemmo.blockus.BlockusListener;
-import net.laboulangerie.laboulangeriemmo.blockus.BlockusRestoration;
 import net.laboulangerie.laboulangeriemmo.commands.Combo;
 import net.laboulangerie.laboulangeriemmo.commands.MmoCommand;
 import net.laboulangerie.laboulangeriemmo.commands.Stats;
 import net.laboulangerie.laboulangeriemmo.commands.TownyMmo;
 import net.laboulangerie.laboulangeriemmo.core.Bar;
 import net.laboulangerie.laboulangeriemmo.core.abilities.AbilitiesDispatcher;
+import net.laboulangerie.laboulangeriemmo.core.blockus.BlockusDataManager;
+import net.laboulangerie.laboulangeriemmo.core.blockus.BlockusListener;
+import net.laboulangerie.laboulangeriemmo.core.blockus.BlockusRestoration;
 import net.laboulangerie.laboulangeriemmo.core.combo.ComboDispatcher;
 import net.laboulangerie.laboulangeriemmo.core.mapleaderboard.LeaderBoardManager;
 import net.laboulangerie.laboulangeriemmo.core.particles.EffectRegistry;
