+++ src/main/java/net/laboulangerie/laboulangeriemmo/commands/MmoCommand.java
@@ -23,8 +23,8 @@ import org.jetbrains.annotations.Nullable;
 import net.laboulangerie.laboulangeriemmo.LaBoulangerieMmo;
 import net.laboulangerie.laboulangeriemmo.api.player.MmoPlayer;
 import net.laboulangerie.laboulangeriemmo.api.talent.Talent;
-import net.laboulangerie.laboulangeriemmo.blockus.Blockus;
-import net.laboulangerie.laboulangeriemmo.blockus.BlockusDataHolder;
+import net.laboulangerie.laboulangeriemmo.core.blockus.Blockus;
+import net.laboulangerie.laboulangeriemmo.core.blockus.BlockusDataHolder;
 import net.laboulangerie.laboulangeriemmo.core.mapleaderboard.LeaderBoardManager;
 
 public class MmoCommand implements CommandExecutor, TabCompleter {
