+++ src/main/java/net/laboulangerie/laboulangeriemmo/core/blockus/BlockusListener.java
@@ -1,4 +1,4 @@
-package net.laboulangerie.laboulangeriemmo.blockus;
+package net.laboulangerie.laboulangeriemmo.core.blockus;
 
 import java.util.List;
