+++ src/main/java/net/laboulangerie/laboulangeriemmo/core/blockus/BlockusOutputStream.java
@@ -1,4 +1,4 @@
-package net.laboulangerie.laboulangeriemmo.blockus;
+package net.laboulangerie.laboulangeriemmo.core.blockus;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
